server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
realesrgan==0.3.0
gfpgan==1.3.8
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
basicsr>=1.4.2
facexlib>=0.3.0
